feat: typeahead has been abstracted into a renderless component and all form module fields now live in a "fields" subdirectory
