+++ packages/fe/components/form/fields/typeahead.vue
@@ -1,121 +1,41 @@
 <template>
-  <div
-    :class="[
-      'field field-typeahead',
-      state, {
-        focused,
-        empty,
-        'dropdown-open': dropdownOpen,
-        'first-option-highlighted': firstOptionHighlighted,
-        'no-results': noOptionsMatchSearch
-      }]">
+  <Typeahead
+    :input-type="scaffold.inputType"
+    :input-placeholder="scaffold.placeholder"
+    :input-autocomplete="scaffold.autocomplete"
+    :input-id="field.fieldKey"
+    :input-disabled="scaffold.disabled"
+    :input-value="field.value"
+    :option-display-key="scaffold.optionDisplayKey"
+    :option-return-key="scaffold.optionReturnKey"
+    :options="options"
+    :aria-labelled-by="modelKey || fieldKey"
+    :class="['field field-typeahead', state]"
+    v-on="$listeners">
 
-    <div v-if="disabled" :class="['input', { disabled }]">
-      {{ value }}
-    </div>
+    <template #option-native-default-text>
+      Select an option
+    </template>
 
-    <div v-else class="input-container">
-      <input
-        :id="fieldKey"
-        v-click-outside="closeDropdown"
-        :type="inputType"
-        :name="fieldKey"
-        :placeholder="placeholder"
-        :value="value"
-        :autocomplete="autocomplete"
-        :class="['input typehead-input', state]"
-        @focus="openDropdown"
-        @blur="closeDropdown"
-        @keydown="handleKeydown"
-        @click="openDropdown"
-        @input="handleInput"
-        v-on="$listeners" />
-      <div v-if="typeof chars === 'number'" class="char-validation">
-        {{ chars }}
-      </div>
-    </div>
+    <template #option-custom="{ option, highlighted, selected, highlightText }">
+      <div
+        :class="['option', { highlighted, selected }]"
+        v-html="highlightText(option)" />
+    </template>
 
-    <Select
-      ref="dropdown"
-      :options="filteredOptions"
-      :aria-labelledby="modelKey || fieldKey"
-      :handle-click-outside="false"
-      :maintain-index-state="false"
-      @dropdownToggled="dropdownToggled"
-      @optionSelected="optionSelected"
-      @optionHighlighted="optionHighlighted">
-
-      <template #option-native-default-text>
-        Select an option
-      </template>
-
-      <template #option-native-text="{ option }">
-        {{ option[optionDisplayKey] }}
-      </template>
-
-      <template #option-custom="{ option, highlighted, selected }">
-        <div
-          :class="['option', { highlighted, selected }]"
-          v-html="highlightText(option)" />
-      </template>
-    </Select>
-
-  </div>
+  </Typeahead>
 </template>
 
 <script>
-/**
- * Here's an example of a typeahead data structure
- *
- * "country": {
- *   "type": "typeahead",
- *   "inputType": "text",
- *   "modelKey": "country",
- *   "label": "Country",
- *   "placeholder": "Country",
- *   "description": "Where are you located?",
- *   "required": true,
- *   "autocomplete": "off",
- *   "pre": "[^\\u0000-\\u00ff]",
- *   "validationMessage": {
- *     "required": "This field is required"
- *   },
- *   "optionDisplayKey": "name",
- *   "optionReturnKey": "code",
- *   "options": [{"name":"Afghanistan","code":"AF"},{"name":"Åland Islands","code":"AX"}]
- * }
- *
- * Note the optionDisplayKey and optionReturnKey.
- * You can pass in any data structure (array of objects) as options and then use
- * optionDisplayKey to define which key in that data structure to match your
- * input text to. Then when an item in the dropdown is clicked, it uses the
- * optionReturnKey to to return the value you want. So for a dataset structure like this:
- *
- * {
- *  "name": "Foo", ← "optionDisplayKey": "name"
- *  "slug": "bar" ← "optionReturnKey": "slug"
- * }
- */
-
 // ===================================================================== Imports
-import Select from '@/modules/form/components/select'
-
-// =================================================================== Functions
-const preValidate = (instance, value, pre) => {
-  if (typeof pre !== 'string') { return }
-  const regex = new RegExp(pre)
-  if (regex.test(value)) { // value contains restricted characters
-    const stripped = value.replace(regex, '')
-    instance.$emit('updateValue', stripped)
-  }
-}
+import Typeahead from '@/modules/form/fields/typeahead'
 
 // ====================================================================== Export
 export default {
   name: 'FieldTypeahead',
 
   components: {
-    Select
+    Typeahead
   },
 
   props: {
@@ -125,156 +45,21 @@ export default {
     }
   },
 
-  data () {
-    return {
-      focused: false,
-      dropdownOpen: false,
-      indexedOptions: [] // hardcode the unfiltered index order
-    }
-  },
-
   computed: {
     scaffold () {
       return this.field.scaffold
     },
-    fieldKey () {
-      return this.field.fieldKey
-    },
     modelKey () {
       return this.scaffold.modelKey
     },
-    label () {
-      return this.scaffold.label
-    },
-    placeholder () {
-      return this.scaffold.placeholder || 'Enter a value...'
-    },
-    inputType () {
-      return this.scaffold.inputType || 'text'
-    },
-    autocomplete () {
-      return this.scaffold.autocomplete
-    },
-    required () {
-      return this.scaffold.required
-    },
-    disabled () {
-      return this.scaffold.disabled
-    },
-    pre () {
-      return this.scaffold.pre
-    },
-    chars () {
-      return this.scaffold.chars
-    },
-    validationMessage () {
-      return this.scaffold.validationMessage
-    },
-    value () {
-      return this.field.value
-    },
-    valueMatchRegExp () {
-      return new RegExp(`(${this.value})`, 'ig')
-    },
-    empty () {
-      const value = this.value
-      return value === undefined || value === ''
-    },
-    state () {
-      return this.field.state
+    fieldKey () {
+      return this.field.fieldKey
     },
     options () {
       return this.scaffold.options
     },
-    filteredOptions () {
-      const options = this.indexedOptions
-      const len = options.length
-      const compiled = []
-      for (let i = 0; i < len; i++) {
-        const option = options[i]
-        const value = option[this.optionDisplayKey]
-        if (value.toLowerCase().includes(this.value.toLowerCase())) {
-          compiled.push(option)
-        }
-      }
-      return compiled
-    },
-    noOptionsMatchSearch () {
-      return this.filteredOptions.length === 0
-    },
-    optionDisplayKey () {
-      return this.scaffold.optionDisplayKey
-    },
-    optionReturnKey () {
-      return this.scaffold.optionReturnKey
-    },
-    firstOptionHighlighted () {
-      return this.highlightedOption === 0
-    }
-  },
-
-  watch: {
-    value (value) {
-      preValidate(this, value, this.pre)
-    }
-  },
-
-  created () {
-    const options = this.options.slice()
-    const len = options.length
-    for (let i = 0; i < len; i++) {
-      options[i].index = i
-    }
-    this.indexedOptions = options
-  },
-
-  mounted () {
-    this.dropdown = this.$refs.dropdown
-  },
-
-  methods: {
-    dropdownToggled (state) {
-      const dropdown = this.$refs.dropdown
-      if (this.dropdown._uid !== dropdown._uid) {
-        this.dropdown = dropdown
-      }
-      this.dropdownOpen = state
-    },
-    openDropdown () {
-      this.dropdown.openDropdown()
-      this.emitToggleFocused(true)
-    },
-    closeDropdown () {
-      this.dropdown.closeDropdown()
-      this.emitToggleFocused(false)
-    },
-    emitToggleFocused (status) {
-      this.focused = status
-      this.$emit('toggleFocused', status)
-    },
-    handleInput (e) {
-      if (!this.dropdownOpen) {
-        this.openDropdown()
-      }
-      this.$emit('updateValue', e.target.value)
-    },
-    handleKeydown (e) {
-      this.dropdown.handleKeyboardNavigation(e)
-    },
-    optionHighlighted (index) {
-      this.highlightedOption = index
-    },
-    optionSelected (index) {
-      if (index !== -1) {
-        this.$emit('optionSelected', this.options[index][this.optionReturnKey])
-        this.$emit('updateValue', this.options[index][this.optionReturnKey])
-      }
-    },
-    highlightText (option) {
-      const inputValue = this.value
-      const optionValue = option[this.optionDisplayKey]
-      if (inputValue === '') { return optionValue }
-      return optionValue.replace(this.valueMatchRegExp, '<span class="highlight">$1</span>')
+    state () {
+      return this.field.state
     }
   }
 }
@@ -284,36 +69,24 @@ export default {
 $height: 2.5rem;
 
 // ///////////////////////////////////////////////////////////////////// General
-.field-typeahead {
+.field-typeahead.typeahead-container {
   height: $height;
   &.dropdown-open {
     @media (hover: hover) {
       &:not(.no-results) {
-        .input {
+        :deep(.input) {
           border-color: transparent;
-        }
-      }
-      &.no-results {
-        :deep(div.select-container) {
-          display: none;
         }
       }
     }
   }
 }
 
-.input-container {
-  position: relative;
+:deep(.input-container) {
   z-index: 10;
 }
 
-.input-container,
-.input {
-  width: 100%;
-  height: 100%;
-}
-
-.input {
+:deep(.input) {
   display: flex;
   flex-direction: row;
   align-items: center;
@@ -345,45 +118,13 @@ $height: 2.5rem;
 }
 
 // //////////////////////////////////////////////////////////////////// Dropdown
-:deep(div.select-container) {
-  display: none;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  z-index: 5;
-}
-
-:deep(div.select-container) {
-  @media (hover: hover) {
-    &.dropdown-open {
-      display: block;
-      select.select {
-        &.native {
-          display: none;
-        }
-      }
-    }
-  }
-}
-
 :deep(div.dropdown) {
   position: absolute;
   top: calc(100% - 2px);
-  left: 0;
-  width: 100%;
   max-height: $height * 5.5;
   background-color: $aztec;
   border: 2px solid $titanWhite;
   border-radius: 0.3125rem;
-}
-
-:deep(div.selection-window-wrapper) {
-  display: none;
-}
-
-.label {
-  z-index: 10;
 }
 
 :deep(div.selection-window) {
@@ -391,9 +132,6 @@ $height: 2.5rem;
   flex-direction: row;
   justify-content: space-between;
   align-items: center;
-  width: 100%;
-  height: 100%;
-  transition: 150ms ease-out;
 }
 
 :deep(div.option) {
@@ -403,7 +141,7 @@ $height: 2.5rem;
     transition: 150ms ease-in;
     background-color: rgba(white, 0.1);
   }
-  span {
+  .highlight {
     font-weight: 700;
   }
 }
