+++ packages/fe/components/form/fields/wysiwyg.vue
@@ -24,7 +24,7 @@
 
 <script>
 // ===================================================================== Imports
-import Wysiwyg from '@/modules/form/components/Wysiwyg'
+import Wysiwyg from '@/modules/form/fields/wysiwyg'
 
 // ====================================================================== Export
 export default {
